Allow override of the repo filename
Apt cannot have 2 mirrors with the same content in 2 different files.
If a deployer has an apt mirror with uca, the deployer
still need to add a repository, but will also need to define the
filename used, in order to avoid clashes.

This commit makes possible to decide the filename for the repo, and
has no master conterpart, because the PPA is only used in Trusty.

Change-Id: Ic6bf8435b06dd866d0c8b984bdbaa7b5fff75e1b