Improve role idempotency
Add a 'creates' argument to the task installing HATop so that the task
is skipped if it's already installed. Remove 'changed_when' from slurp
tasks. Break the 'Test for log directory or link' shell task into a
couple tasks which use the stat and file module. Remove 'changed_when'
from slurp tasks. Also enable the idempotence test for the gate.

Partial-Bug: 1640144
Change-Id: Id7fe24c9c1ae416abab69d4f42b65c9594805e10