Remove unused/unnecessary haproxy_server vars/files
The repo/keys are left over from Newton/Trusty and
were not removed when Trusty support was removed.

The required packages were only necessary in order
to facilitate the repo addition.

The var haproxy_distro_packages is defined in all
distro-specific vars files, so its presence in
defaults is unnecessary.

The apt pinning meta dependency is no longer
required - it's another leftover from Ubuntu Trusty.
A task is included to remove the old config files.

Change-Id: I912cd170d05c4a9befe3420971ddf68ff2ddde2b