Add option to set the check port
The option to set the checkport is needed when trying to balance traffic
to a service that uses a secondary port to verify service availability.
This is being implemented to support galera which should be using an
HTTP based check to verify sync status instead of a simple mysql login.
For reference this is what the folks at Percona are recommending [0].

[0] https://www.percona.com/doc/percona-xtradb-cluster/LATEST/howtos/virt_sandbox.html
Closes-Bug: #1665667

Change-Id: I81c924464aa4b19c2a62f37b5bf26c3c0453786a
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>