Add ability to manage non-inventory hosts
We may want to load balance to existing services that we aren't managing
via ansible. Currently the hosts have to exist in the inventory in order
to add a VIP for these hosts, this patch adds the ability to set the
hostnames and addresses of the hosts manually when they aren't in the
ansible inventory.

Additionally, this patch adds a test for both the group method and the
host_lists method.

Change-Id: Ida66f401d8320d9bf14eac9b8014124631978808
(cherry picked from commit d2b2492c16711dbf1971ef513bcd43d33339cf38)