Fix use of haproxy_backend_ssl when haproxy_backend_ca is not defined
For certificates from widely trusted CA there is no need to provide
a specific CA file for an ssl backend, but the code fails with
undefined variable if only haproxy_backend_ssl is enabled.

A workaround is to set `haproxy_backend_ca: false` but this patch
allows haproxy_backend_ssl to be used on it's own.

Change-Id: I7c87317433acb4ed73070a2252240737b22dccfc