Fix haproxy_stats SSL path defenition
Neither `vip_interface` nor `vip_address` are defined or available in
the context they're being used.
Thus we need to refer to available variables in order to render base config
properly

Current version fail with "AnsibleUndefinedVariable: 'vip_interface' is undefined"
on "Drop base haproxy config" task.

This fix the issue that was introduced with [1] and backported back to Zed

[1] https://review.opendev.org/q/Ib8be6b7fc3dada9d20905b0f07d90ddce0335605

Change-Id: I4e52378d8c5b3eaa6863ecaf0d04554d082e3dc0