Generate PKI SAN as a list
This transforms generation process from concatenating strings to a
list which is joined before passing to the PKI role.

It improves code redability and is more error-prone as we don't need to
pay close attention to presence of `,` separator while concatenating
strings.

It will also allow to pass only unique records to PKI role if we have a
need for that in the future.

Change-Id: I88f74c97592795704170b49bff56b23bc8145f5b