Add ability to extend TLS SAN per VIP
At the moment it is really tricky to extend list of domains for which
TLS certificates should be generated.
While this might be required for some cases, like domain-based endpoints

While this options does not provide opprotunity to pass user-created
certificates, it enables to leverage PKI role for generating ones.

Needed-By: https://review.opendev.org/c/openstack/openstack-ansible/+/934536
Change-Id: I80691ad5e6b51181db44420b3b33967184589b68