Rename haproxy_tls_vip_binds to haproxy_vip_binds
Effectively haproxy_tls_vip_binds was influencing not only TLS-related
binds but also non-TLS ones. It's naming was picked in a misleading way
from start, as a lot of logic for the variable was around TLS
implementation. Though it has a global effect and this should be
reflected in it's name.

Change-Id: I2f28d0ab033eff96cc2bc09db917bedb387013e8